refactor : 테스트 코드에 프로덕션 코드의 패키지 구조를 도입
