+++ src/test/java/domain/WinningStatusTest.java
@@ -1,6 +1,7 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.WinningStatus;
 import org.junit.jupiter.api.Test;
 
 public class WinningStatusTest {
